Default types, threadsafe module, better docs
There is a new threadsafe module that exports aliases that make using
this crate in a threadsafe crate much simpler. Default generic types are
also used to make it so that the root types can be used with minimal
generic parameters to automatically operate in single-threaded mode.

Mode has been renamed to Behavior, as the implementation wasn't really
the modal part of the crate.

get_unchecked and get_mut_unchecked have been added as variants that use
debug_assert instead of assert. This allows well-tested code to skip all
runtime checks when compiled in release mode, while being fully verified
when debug_assertions are enabled. These new functions are unsafe due to
the potential for undefined behavior if code was never tested with
debug_assertions enabled.